appsec_sprint_evaluator/pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "appsec_sprint_evaluator"
version = "0.1.0"
description = "A CLI tool to evaluate Application Security (AppSec) tooling and integrate them into the SDLC for Project Jupyter."
dependencies = [
    "requests",
    "PyGithub",
    "google-generativeai",
    "colorama"
]
requires-python = ">=3.8"

[tool.setuptools]
packages = ["appsec_sprint_evaluator"]

[project.scripts]
appsec-eval = "appsec_sprint_evaluator.cli:main"
appsec-tutorial = "appsec_sprint_evaluator.tutorial:main"
